Moved all project listing component classes to new folder.
